refactor: move current xwin implementation to `clang_cl` backend
