+++ src/lib.rs
@@ -1,3 +1,4 @@
+mod backend;
 mod common;
 mod macros;
 mod run;
